shahia_backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
flask==3.1.1
flask-cors==5.0.0
flask-sqlalchemy==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.10.0
markupsafe==3.0.2
openai==1.98.0
pydantic==1.10.17
pydantic-core==2.14.6
PyJWT==2.10.1
requests==2.32.4
sniffio==1.3.1
sqlalchemy==2.0.36
tqdm==4.67.1
typing-extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.5.0
werkzeug==3.1.3

